Avoid spilling arguments for function calls
Summary:
There was a crash in a react native snapshot test when we attempted to
turn on static builtin by default. The assertion failure was in `ISel::verifyCall` and
the arguments were misplaced for a `CallBuiltin` instruction. It turns out we need
to avoid moving arguments to function calls to spill registers because the calls
assume the arguments are placed at the end of the stack. Add this check for `CallBuiltin`
and `CallDirect`.

Reviewed By: tmikov

Differential Revision: D14249891

fbshipit-source-id: fbeba3290b36222f9f47d421a9215207d5da0eeb