Throw an exception on attempt of overriding static builtins
Summary:
Add a flag to the property descriptor to indicate a property is a static builtin object/method.
When we freeze the builtin on loading static-builtin-enabled bytecode, we set the flag
for those builtins. And when there is an attempt to override the builtin, throw an exception,
even if it's in non-strict mode, when we normally don't throw an exception.
This makes it easier to catch source code errors caused by overriding builtins.

Reviewed By: tmikov

Differential Revision: D13629671

fbshipit-source-id: 853d058b121d2f7077b44c0d4405d4cecf3ba114