Hard limit the max register stack size in RuntimeConfig
Summary:
Prior to this fix, there was no limit on the number of registers that
could be requested in a RuntimeConfig. In particular the number of
registers could be so large that the allocation size
`count * sizeof(PinnedHermesValue)` would overflow, and we would
certainly crash.

Add a hard limit kMaxSupportedNumRegisters = UINT32_MAX /
sizeof(PinnedHermesValue), which is about 536 million registers.

This assists with the prior commit, which requires a UINT32_MAX register
count to be invalid.

Reviewed By: tmikov

Differential Revision: D14211536

fbshipit-source-id: fd32aa89f8d09436b3abe8c4e0af7f8972ca621b