Be more careful about overflow when malloc()ing
Summary:
Hermes has a function checkedMalloc() which takes a total size. We routinely
compute this size by multiplying a count by an element size. However we are
sloppy about checking that this calculation does not overflow.

Introduce a new function checkedMalloc2() which takes a count and a size,
and fails on overflow.

Switch checkedMalloc* to inline. This allows constants to be propagated
into the checkedMalloc2() overflow check, allowing e.g. shifts instead
of division.

Switch every call site of the proper form to checkedMalloc2().

Remove checkedCalloc(), it was unused.

Reviewed By: tmikov

Differential Revision: D14211534

fbshipit-source-id: d5c9c98e25875fd9020c3b1f77930c6605053512