Guard against stack overflow when computing outgoing registers
Summary:
Prior to this fix, when constructing a CallFrame, if the number of
arguments was very large, then caller outgoing register count calculation
would overflow to a small value, and we would not correctly detect stack
overflow.

Fix this by saturating the callerOutgoingRegisters calculation,
effectively turning UINT32_MAX into a sentinel value. This means that on
(say) UINT32_MAX argument count, we will return UINT32_MAX when the
register count ought to logically be UINT32_MAX+7. However this difference
is irrelevant, because both sizes are larger than any supported stack size.

A followup commit adds assertions that prevent the supported stack size
from being so large, to enforce this constraint.

Reviewed By: tmikov

Differential Revision: D14211535

fbshipit-source-id: 5f2ff210c81fece7878c467bac47d6f442511957