Make the SourceErrorManager messageCount_ array the proper (new) size.
Summary: The change to the new LLVM introduced a bug that was crashing some apps.  A clue was that it was only happening in dev situations, with the Metro bundler running, so the app was running from JS source, rather than bytecode.  Turns out that an enum defined in LLVM gained an extra value, and an array in Hermes, indexed by the int values of this enum, only had enough entries for the old size.  So there would be a buffer overrun in the compiler.  The fix is quite easy.

Reviewed By: tmikov

Differential Revision: D14294568

fbshipit-source-id: 585fb45ebcf2f52149461f51e30511e8f8b66f1f