Add equivalent year logic in ES5.1 15.9.1.8
Summary:
This is necessary because some library calls doesn't work when the input date-time cannot be represented as a 32-bit non-negative number of seconds since 1970-01-01T00:00:00. (e.g. std::localtime on Windows)

Without this change, constructing date object that is before 1970 or after 2037 fails on Windows. This logic is applied to all platforms in order to have consistent behavior across platforms.

Reviewed By: dulinriley

Differential Revision: D14196805

fbshipit-source-id: 1ecce8b3de2916ae0ffa2ca746054e829f3972ec