Add support for getPropertyNames to API tracing and replay
Summary:
There was a bug introduced to tracing by the jsi decorator stack; namely, handling of `getPropertyNames`.
The current implementation of `getPropertyNames` calls `createArray` and `setValueAtIndex`.
Before the decorator stack, this would be forwarded to a tracing function, and everything would work out fine.
After the decorator stack, these functions are *not* traced, and arrays seemingly come from nowhere.

The solution is simple: instead of relying on implementation details of `HermesRuntimeImpl`, instead
explicitly capture and replay the `GetPropertyNames` event.

Reviewed By: amnn

Differential Revision: D14287687

fbshipit-source-id: 65e4ee13bcc3078920aa93b25c0ad5502e5ff8bf