Fix handle count assertion in HermesInternal.getInstrumentedStats.
Summary:
There were enough `SET_PROP` calls that we were running out of handles
in `HermesInternal.getInstrumentedStats` in certain conditions.

`defineNewOwnProperty` may allocate some handles, and when we're doing
this many times in a row, we can simply flush the scope every time we
call it to ensure things don't get out of hand.

Reviewed By: tmikov

Differential Revision: D14387895

fbshipit-source-id: 531aaec454fd642aaf1ce6154d502df91ce2dbee