Fix handle-san problem in regex getSubstitution
Summary:
There was a use of a raw pointer via the StringView iterator `itr` in `getSubstitution`.
When the lambda `submatchOrEmpty` calls `createStringView`, it might (in the future) cause
an allocation from any Rope-like data structure. So it forces collections to catch bugs.

The fix is to not store an iterator which caches the pointer value, and instead fetch it fresh each
time from `replacementView`.

Reviewed By: willholen, avp

Differential Revision: D14387951

fbshipit-source-id: 8135164288ed516390ce2a3f1dfc60fc20f13787