Separating Predefined Symbols from Predefined Strings
Summary:
In order to improve IdentifierTable initialisation I/O, I need to add a
constraint to the predefined strings list that it contains no duplicates.  This
constraint does not apply to Predefined Symbols.  Additionally, as they are
currently used, we typically include `PredefinedStrings.def` twice, once with
with `STR` defined and once with `SYM` defined, which seems to support the idea
that they may belong in separate files.  There is currently one situation where
that is not true -- the definition of `Predefined` in `Runtime.h`, but in order
to guarantee initialisation order matches the order of enums in `Predefined`
(another assumption that needs to be formalised for `IdentifierTable` init I/O)
I would need to change that also (I have done so as part of this diff).

Reviewed By: avp

Differential Revision: D14242735

fbshipit-source-id: 3a6beab773930826735a1f9877598e8a9038bee2