Merge ReservedSymbolID::(empty|deleted), SymbolID::create(Empty|Deleted)()
Summary:
`ReservedSymbolID::empty` and `SymbolID::createEmpty()` (and `::deleted` and
`::createDeleted()` respectively) are used interchangeably throughout the
codebase.  This diff merges them into a single constexpr function on
`SymbolID`.

I did this for the following reasons:

 - `ReservedSymbolIDs.h` currently pertains only to `InternalProperty` slots.
   As I will be introducing a related concept under a similar name, I wanted to
   focus the naming here to be specific to `InternalProperty`.
 - The `Empty` and `Deleted` SymbolIDs are relevant in the context of all
   SymbolIDs, not just the ones used by `InternalProperty` slots, so before
   making the name change, I wanted to move them out of there.
 - When ramping up on the codebase, I had found the distinction between these
   two concepts (and the fact that they were defined in different scopes) a
   little confusing.

As `SymbolID` is used as a value type, I chose to elide the `create` from the
constexpr name during the merge.

Ideally I would have used a `constexpr` member field on `SymbolID` for this
purpose (as opposed to a function) but details on how the C++ spec requires
`constexpr` member fields to be intialised prevents this.

I tried to use the suggestions in [this Stack Overflow](https://stackoverflow.com/questions/11928089/static-constexpr-member-of-same-type-as-class-being-defined):

```
class SymbolID {
  static const SymbolID Empty;
};

/* static */ constexpr SymbolID Empty = SymbolID{EMPTY_ID};
```

Which compiles, but does not link, due to duplicate symbol errors.  My theory
is that the `constexpr` at the definition messes up the default internal
linkage implied by the `static const` in the declaration.  Adding a redundant
`const` at the definition also did not help.  I gave up on that there, opting
to highlight the issue in a spec comment.

Reviewed By: avp

Differential Revision: D14303554

fbshipit-source-id: c6506a6219414a027e9279d8bfb4c3f9f756ba0d