Move initialisation of Reserved Symbol IDs into initPredefinedStrings
Summary:
Previously this happened in `IdentifierTable`'s constructor.  I am moving it
into `Runtime::initPredefinedStrings` to keep all the code that cares about and
affects the initial order of SymbolID allocations in one place.

I also have plans to coallesce these calls with the calls to reserve SymbolIDs
for the predefined symbols.

Reviewed By: avp

Differential Revision: D14260595

fbshipit-source-id: 3f3736c812c19423dcef98b4c79827f14ce54e44