Deduplicate Predefined Strings due to CellKinds.
Summary:
Predefined Strings generated due to CellKinds were introducing some duplicate
strings.  In moving to a world where the Compiler and Runtime can statically
agree on SymbolIDs for Predefined Strings, it becomes useful to ensure they
are unique.

Duplicates came about because
 - Each CellKind needs an associated JS Class Name
 - Multiple CellKinds have no such name (they are internal cell types used only
   by the VM), so were providing an empty string
 - Some CellKinds map to the same JS Class name.

Predefined Strings due to CellKinds were defined by mapping over every kind and
extracting the provided JS Class Name.  But this means that kinds with no good
name result in duplicate strings, as do kinds that map to the same name as each
other (Object, HostObject and various flavours of Function).

This diff decouples `CellKind`s from their JS Class Name, allowing the former to be defined without the latter, or for multiple kinds to map to the same class name.  There are now 3 macros in play in this file:

 - `CELL_KIND`: Just used to define the kind.
 - `CELL_JS_NAME`: Just used to associate a class name with a CellKind.
 - `CELL_CLASS`: A convenience macro for defining a kind in 1:1 correspondence with a JS Class Name.

The majority of kinds can be mapped to a single JS Class, so this does introduce some churn in this diff, renaming them to `CELL_CLASS`, but this felt like the most appropriate naming scheme given the concepts at play here.

Reviewed By: avp

Differential Revision: D14242736

fbshipit-source-id: 17288de4ba2e9b1a87acf7689d8c7df9d4f0ccd9