Rename SymbolID::is{External,Internal} -> SymbolID::is{NotUniqued,Uniqued}
Summary:
As part of my work to make the Predefined String enum values correspond exactly
with their SymbolIDs, it becomes convenient to have what are currently referred
to as the "Reserved Symbols" (the Internal Property slots) have Predefined IDs.
This moves the creation of their corresponding SymbolIDs outside of
`IdentifierTable`, necessitating a public facing API.

Hermes currently has two concepts -- External Symbol IDs and Reserved Symbol
IDs -- which look the same from the IdentifierTable's point of view.  They both
require the addition of a SymbolID without it being accessible via the
`IdentifierHashTable` however, they achieve this through two different APIs.

The API used by External Symbols almost fits the bill, except that its
specification makes explicit reference to External Symbols and their
association with JS' Symbol abstraction.  This diff removes mention of that
from the spec and function names so that it could be used for both purposes.

There was another naming issue in that `isExternal` conflicts with the notion
of `ExternalStringPrimitive` and `isInternal` conflicts with `InternalProperty`
slots -- both of these concepts are semantically quite close to `SymbolIDs`,
introducing opportunities for confusion.

The renaming was done in quite a brain dead way (almost a straight find and
replace from `External` to `NotUniqued`, and correspondingly for `Internal` and
`Uniqued`), so there may be better ways to reword function names etc.

Thanks to avp for helping me understand what this flag means and find a better
name for it.

Reviewed By: avp

Differential Revision: D14244326

fbshipit-source-id: c29fec5663251a11c6eefa5896747766b916d93d