Initialise Predefined Strings before special Runtime Module
Summary:
Creating a bytecode-backed RuntimeModule always imports at least one string (if
there are no strings in the module, one will be added).  I'm not sure why this
is, but because of this and the fact that the "Special CodeBlock RuntimeModule"
was previously created before the predefined strings were initialised, its
"filler" string was the first entry in the `IdentifierTable`.

Moving predefined string initialisation before this RuntimeModule is created to
sidestep this issue for now. (We may have to revisit when adding support for
user-defined reserved strings).

Reviewed By: kodafb

Differential Revision: D14259655

fbshipit-source-id: f9581c673a476381ed96e4fec1e41344826e001b