Split Predefined enum into IProp, Str, Sym
Summary:
The SymbolIDs for Internal Property Slots and Symbols are not uniqued, whereas
the IDs for Strings are.  The aim of the next diff in the stack is to make it
possible to translate from the predefined enum to its SymbolID statically.
Because the SymbolID indicates whether the identifier is uniqued or not, we
need to be able to statically differentiate between uniqued and non-uniqued
predefined identifiers.

A previous attempt at this grouped property slots and symbols next to each
other (in terms of the predefined enum values they occupied) so that there were
only two kinds of Predefined identifier: `Uniqued` and `NotUniqued`.  But we
found this measurably regressed performance (we suspect by making property
lookups by SymbolID less efficient by requiring more hash table probes).  This
diff opts to split `Predefined` into three instead, so that order can be
preserved whilst making it possible to statically derive the Predefined ->
SymbolID mapping.

Reviewed By: avp

Differential Revision: D14279441

fbshipit-source-id: 98fb2b97387b99e74a59ab5759cf20476d8030bc