Moving Predefined enum into its own header.
Summary:
Necessary in the long term so that the compiler and runtime can both have
access to this enum.  In the shorter term, it makes some of the consolidation
work between ReservedSymbolIDs and PredefinedSymbols slightly easier.

This file will need to be moved (and perhaps have its namespace changed) when
it is eventually referenced by the compiler.  I'm deferring that change until
it becomes necessary -- so far, it remains only referenced from within the VM
(although soon from multiple locations).

Reviewed By: avp

Differential Revision: D14287405

fbshipit-source-id: 8efab098b671d716fac94465bf7048840b5b88b2