Introduce Domain.
Summary:
Add `Domain`, an object which we will use to GC-manage RuntimeModules.
JSFunctions maintain strong references to `Domain`s,
and when `Domain`s are destroyed, their finalizers will remove
any `RuntimeModule`s will be `delete`d.

This method of managing `RuntimeModule` lifetime replaces the refcount
that we used previously, so delete the `addUser` and `removeUser`
code that does that.

For now, we leave the list of `RuntimeModules` in the `Runtime`
to avoid breaking other parts of the VM.
That list may be removed in a later diff.
We also use `std::vector` to avoid having to throw inside
RuntimeModule creation upon addition to an `ArrayStorage`.
This may be changed in the future if we have many Domains.

This is the first step to fixing a memory leak when using CommonJS
modules; the module table was keeping all the RuntimeModules alive,
and moving the CommonJS table to the Domain will allow collection
because the Domain is in the GC heap.

Currently, because we only support running with a single Domain,
there are no functionality changes.

Reviewed By: tmikov

Differential Revision: D13540656

fbshipit-source-id: 6c2e40388fa8b1620db4279839dd56502b3d9682