Move CJS modules to Domain.
Summary:
For the next step toward bundle splitting, we need to centralize
all of a Domain's modules in the Domain itself.

To do this, we remove the CJS module storage from the RuntimeModule
and make it GC-managed, allowing us to place it in the Domain.

This has the side-effect of requiring RuntimeModule initialization
to be able to throw, as pushing onto or creating an ArrayStorage can
still throw an exception.

To avoid having to deal with this issue (as it only occurs when using the
CommonJS module table), introduce `initializeWithoutCJSModules` to allow
users to initialize `RuntimeModule` without having to deal with the
`ExecutionStatus`. This means we don't have to ignore any failures,
we simply need to be conscious of whether the new RuntimeModule can
contain a CommonJS module table.

As of right now, we have no mechanism to import multiple CommonJS module
tables into the `Domain`, but that will change shortly.

Reviewed By: tmikov

Differential Revision: D13577147

fbshipit-source-id: 9bc9250864b7c139b35091fcb5938c95da3dafb0