Split bundles based on segments.
Summary:
This lets Hermes actually emit multiple bundles.  The files emitted are
simply the actual output filename with `.n` appended to the end, where
`n` is the ID of a given segment (as specified in `metadata.json`).

This is accomplished by passing the range to emit along the BC generation
path, and skipping generation for any functions that aren't in the
segment. Determining whether a function is in the segment is done based
on walking up the lexical scope chain until we find a parent which has a
CJS module ID.

Reviewed By: tmikov

Differential Revision: D13869186

fbshipit-source-id: 32e2bc12f65ec837f5dabd12b31a9dcf73a99041