Introduce RequireContext and simplify module entry.
Summary:
We eventually need to pass the Domain around as we call `require`,
so enable this by binding a proper new object `RequireContext` to the
`require` function instead of just the directory name.

As a result of `RequireContext` requiring a `Domain` to create,
we eliminate the JS `HermesInternal.require` call and add a check in
`Runtime::runBytecode` to pass the `Domain` along and call the first
module directly (the `Domain`, unlike the initial directory name, is not
always the same at compilation time).

`RequireContext` must be a `JSObject` because it will eventually be
exposed to users through `require.context` and will need to be passed
around as a `jsi::Value`. While users shouldn't modify it, the VM
should not assert if there is an attempt to modify or print it.

Reviewed By: tmikov

Differential Revision: D13869173

fbshipit-source-id: 8c9ff01b1efb4056874f2f68feaab41b83e747d4