Allow loading properties from require() in static resolution.
Summary:
We need to be able to load `require.context` regardless of whether
we have statically resolved the `require` expressions, because the
user must be able to use `require.context` to load new segments.

Allow loading properties from `require.context`, which we can do
because storing properties still will cause a fallback, so there must
not be a getter on `require.context`.

Reviewed By: tmikov

Differential Revision: D14094236

fbshipit-source-id: 20a72319909972f287c30d696ebf04bfc9220e82