Add loadSegment to ConsoleHost.
Summary:
Add a `loadSegment` function to be used to test segment loading.
This requires passing the base bytecode file name through,
allowing the function to read segment files from disk.

Because this is simply for console, the function simply appends `.n`
as a suffix to the base filename, where `n` is the ID of the segment.

Reviewed By: tmikov

Differential Revision: D13869317

fbshipit-source-id: e72d97809d57e8af6fc4279895d293e8b936fa31