Introduce new CrashManager interface and use it for the Runtime and registers
Summary:
Add a new `CrashManager` interface type (injected into `RuntimeConfig`) that can be used to register important pieces of memory to be included in crash dumps outside of the stack.

For every `Runtime` constructed, put it into the "app memory" by registering it with the `CrashManager`. Un-register it in the destructor.

Reviewed By: davedets

Differential Revision: D13616142

fbshipit-source-id: 4a591deef5368bf2b2259cfc880401a6ae2e9433