Add OSS cross-compiling support to cmake_build.sh
Summary:
This allows cmake_build to do the OSS build after the regular CMake build.

The alternative approach would be to do both separately, but this would
require two jobs that both need to build LLVM which seems like a waste.

On Sandcastle, this is currently forced to make a writeable copy of the SDK. I've already posted to Android SDK/NDK Upgrades to see if they can accept the licenses so that this is no longer necessary: https://fb.workplace.com/groups/549570248562485/permalink/1044006889118816/

Reviewed By: haozhun

Differential Revision: D14325197

fbshipit-source-id: 431e0186f7d69ac160deaf6a8c807c446ced719a