Add option for specifying the sanitizer seed to the GC
Summary:
Add a command line flag to pass to the GC to request a specific seed be used
for handle sanitization.
This allows an easy repro when a specific seed causes an error.

This is *especially* useful when there's a bug in sanitization and changing the rate to 1
does not expose the bug.

Reviewed By: davedets

Differential Revision: D14378904

fbshipit-source-id: a0048d8fe58b525f30c6acd66d1d943e2f20845f