Change StringPrimitive::ensureFlat to do an alloc instead of a collection
Summary:
`StringPrimitive::ensureFlat` queried the GC to ask if handle-sanitization was on, then forced
a collection if that was true.
This is not the same semantics as a possible allocation for all GCs, so it is better to instead
do a dummy allocation. The GC will detect if it needs to sanitize handles at that point.

Note that this means with the current NCGen (which does nothing on handlesan), this will only have a small
possibility of causing a collection.

Reviewed By: amnn

Differential Revision: D14388248

fbshipit-source-id: 440382551f170fe956079e9cc731e816a78df849