Change DummyRuntime to use a proper StorageProvider setup
Summary:
`DummyRuntime` was not using the correct `StorageProvider` setup.
In a world where pointers can be compressed, it's very important to get the correct
`StorageProvider`.
Some customization is allowed for decorators, but use that API with caution.
Choosing the wrong `StorageProvider` can cause bugs in the VM and GC.

Add `DummyRuntime::defaultProvider` to make the choice clear for users.

Reviewed By: ridiculousfish

Differential Revision: D14433408

fbshipit-source-id: b61864c04028e63ebe04ad9f0c379c02eda2b1be