Add cpu-cycles to perf event counters
Summary:
Add counter for CPU cycles. The number of cycles per second can vary over time if frequency scaling and similar is active (in a benchmarking environment, it shouldn't be).

This metric correlates very strongly with totalCPUTime and totalTime in our benchmarks, but has less noise. (If it does not correlate well, it suggests CPU frequency is not constant.)

Reviewed By: amnn

Differential Revision: D14427898

fbshipit-source-id: c35e58daf91a6c848fd1fb3f5ad48c090521014a