Remove ReservedSymbolID marking from IdentifierTable.
Summary:
The code checked whether the reserved symbol was a string primitive and then
marked it.  All the current reserved symbols have no useful string
representation, and should not be materialised, so this code was not being
used.  I am going to limit the reserved symbol IDs to be used only for internal
property slots so I expect it to never be true, so am removing this code.

This also removes the last reference to `ReservedSymbolID` from within
`IdentifierTable`, completing the separation of concerns.

Reviewed By: kodafb

Differential Revision: D14323528

fbshipit-source-id: 1735ce1eaba780699fb00d4a71b3be0dbd9ed123