Move getPredefinedSymbolID into Predefined.h
Summary:
It was on `Runtime` before because it needed access to the `predefinedStrings_`
lookup vector.  This is no longer the case and the signature has been updated
to show that it is a static constexpr function, but call-sites still treated it
like a member function.  A side-effect of this change is that this is no longer
the case.

The main reasons for doing this are that

 - In the short term, `InternalProperty` needs access to this functionality
   (and currently duplicates it).
 - In the longer term, the compiler will need access to this functionality
   also.

Reviewed By: kodafb

Differential Revision: D14327192

fbshipit-source-id: 13ee1794d5983721740d1fcb146616bf100b5d8e