Introducing `isSymbolPrimitive` and `isPropertyNamePrimitive`.
Summary:
Make the relationship between not uniqued SymbolIDs, JS Symbol instances and
Internal Property slots more explicit in the form of two helper predicates:
`isSymbolPrimitive` and `isPropertyNamePrimitive` which indicate whether the
SymbolID passed in is backing a JS Symbol or a named property.  Used in the
implementations of `getOwnPropertyNames` and `getOwnPropertySymbols` in
`JSObject`.

This diff also corrects a few typos etc. near the code I touched.

Reviewed By: tmikov

Differential Revision: D14330824

fbshipit-source-id: 5e2d126367e1a0ff379759aa76473692aeaaad59