Fix a leak: JSWeakMapImpl<C> was not using the HasFinalizer::Yes alloc function.
Summary:
ASAN detected a leak within the VM.  Looking at it, it was a case of the allocating space for an object with a (GC) finalizer, without specifying in the allocation call that it needed to be recorded as a finalizable object.  JSWeakMap is the type.

This diff fixes the particular problem.  The next diff in the stack puts in a debug check that identified the problem in our normal testing.

Reviewed By: tmikov

Differential Revision: D14452018

fbshipit-source-id: 3b5adac25a88edd315dad742a72c14b632637fb3