Improve Handle-SAN warning messages
Summary:
The error message for handle sanitization was in two different places and didn't give the best
advice for quickly repro-ing crashes.
The new error message has all the information and instructions for a quick repro.

Also, moved `GCBase` constructor into a cpp file. It is too complex to remain in the header.

Reviewed By: avp

Differential Revision: D14506284

fbshipit-source-id: 5c6808d7619e33503f3efe2625fd12ceb495010a