Add checking in dbg that objects with finalizers are allocated with the right alloc function.
Summary:
The previous diff in this stack fixes a leak discovered by ASAN.  There have been other instances of this general problem: if a type has a finalizer, we need to allocate instances of it with an allocation function that notes this, and puts the cell on the finalizer list.  This is easy not to do.

This diff adds checking that detects such errors.  The allocation functions of the generational GCs (GenGC and NCGen) append a cell declared to be finalizable at the end of the finalizer list of the generation they're allocated in.  So the strategy is: in the GCCell ctor, we have the vtable, so we can see if the type has a finalizer.  If so, the "this" must be on one of these finalizer lists.

MallocGC does not keep an ordered list of finalizable objects, so it can't do this checking.  It implements the new API, but in a way where it just doesn't complain.

Reviewed By: dulinriley

Differential Revision: D14460658

fbshipit-source-id: ba9c647de713a61dc5c7953c1ee6570fefd9226d