Add crash callback feature to CrashManager
Summary:
Adds the ability to register callbacks to a CrashManager instance. These will be invoked when a crash happens, allowing us to dump out data we can salvage after a crash. The data is provided via a pre-opened fd, and is intended to be JSON-compliant text output.

Facebook
We provide an implementation of the above API using our internal Breakpad implementation. Crash data is stored in the minidump output and can be extracted with tools like `minidumpaux.py` using the stream identifier MD_FB_HERMES_CRASH_DATA.

To me it makes sense to have one root handler and thus produce at most one MD_FB_HERMES_CRASH_DATA stream output. So, I've set things up so multiple handlers can be registered but are then serially written to one stream. Data managing this at a process-global level are in BreakpadCrashManager.cpp.

Thanks to dulinriley for the first cut of this implementation.

Reviewed By: ridiculousfish

Differential Revision: D14494263

fbshipit-source-id: 64f7825820ca1b3471b6652762295b50ab8bc167