Add CrashManager callback to Runtime which dumps a JS stack-trace
Summary:
Runtimes now register/de-register a CrashManager callback on construction/destruction which dumps out a stack-trace. This is built on the hope that in many cases most data in a Runtime will be intact on crash.

Facebook
We'll use this to help track down the issues associated with S174325 rather than the specific hack introduced in D14418618.

Even if further crashes are caused by this new handler during a crash, my understanding is Breakpad manages this and will just discard this stream.

Reviewed By: ridiculousfish

Differential Revision: D14494260

fbshipit-source-id: 29c5805dcda0eab6ce27b445e0821b7f2047bdaf