Dump AST in JSON format
Summary:
Support dumping AST in JSON format. Replace the old `ESTreeDumper` with
`ESTreeJSONDumper`. So now to dump Hermes AST, the options are:
`-dump-ast` or `-dump-transformed-ast` which will dump JSON on one line;
adding `--pretty-json` will dump the pretty JSON.

We already have the `RecursiveVisitor` for traversing the AST, but it only visits
the values so it can't be easily adapted. To emit full JSON, we will need
 to capture the keys for the JSON object as well.

Reviewed By: tmikov

Differential Revision: D14416591

fbshipit-source-id: 49a25aa66680eeba5d52b109d282927a4fa6b519