Implement String.raw.
Summary:
String.raw is intended to be used as a tag function with template literals.
It produces a string value that is the concatenation of raw string elements
and evaluated substitutions. For example `` String.raw`hello\n` `` becomes
`hello\n` without processing the escape sequence.

Took over from Aakash and added some tests to use it as a tag function.

Reviewed By: tmikov

Differential Revision: D13642104

fbshipit-source-id: f90fd571c71634ce57a270ef97aeda90d38174cf