Add template object cache
Summary:
Implement "templateMap" in the ES6 spec.
It is a cache for template objects: the tagged literals with the same list of
string elements should have identical template objects, so the cache avoids
creating unnecessary template objects. Each runtime module will have such
a `templateMap`.

During runtime, the map maps from numeric ids to pointers template objects.
We determine the ids during compile time, because all the template objects
can be know in compile time.

In compile time, template objects has a one-to-one mapping from the
list of raw strings in the template literals. We maintain such a map and assign
a unique id to each of them.

In the spec, the cache is defined in the "realm", while we have a templateMap
in each runtime module. They are not exactly semantically equivalent. For
example, the template object cache should still be accessible from within `eval`,
and functions created using the `Function` constructor. But we create a new
runtime module in those cases. We won't behave exactly according to the
spec in these cases.

Reviewed By: tmikov

Differential Revision: D14112841

fbshipit-source-id: d7ceb4330466ecff7413f0c24db1a9e2b4a80178