Add synth tracing for evaluateJavaScript() return value
Summary:
The return value is now stored as part of the EndExecJS record. When replaying the trace we assert the actual evaluation result matches the one in the trace.

This probably won't be that useful until we support traces which allow multiple evaluations. At least it's here though, and maybe the result assertion will catch something one day.

Reviewed By: dulinriley

Differential Revision: D14118185

fbshipit-source-id: a5777b14fd40599210f5d72f8b5df74b92e75eb6