Add a magic word to VTable header to catch bugs in prod
Summary:
Adding a magic word to the start of a VTable can prevent confusion if
it is accidentally used as a forwarding pointer (and vice-versa).
Unlike the magic word in GCCell, this magic word can be left on in production
with minimal impact, since VTables are allocated in static data segments in small
numbers.

Add a new `isValid` method for VTables that checks this magic field. `GCCell::isValid` calls
this function.

Differential Revision: D14578991

fbshipit-source-id: 2efee7937529f7e2a193839c10c7dfca70639b4b