Set sources properly for CJS modules in SourceMapGenerator.
Summary:
When setting the `sources` field of the `SourceMapGenerator`,
we must account for the fact that the user may have input a directory
or a zip file, in which case the `cl::InputFiles` field is not the
same as the list of actual JS sources we use.

Account for this by repopulating the `sources` list in
`generateIRForSourcesAsCJSModules` to ensure the order is correct.

Reviewed By: ridiculousfish

Differential Revision: D14530425

fbshipit-source-id: 8f767cc3bbcd1ed19a48022a3d889ec0d8f5fca1