Stop storing string table indices in SourceMapGenerator.
Summary:
SourceMapGenerator used to keep a mapping from the filename table indices
to the `sources` list. This is a relic from before we split the file names
into the filename table.

Source mapping should not be dependent on a `DebugInfo` backing it,
so we use a `std::string -> uint32_t` mapping instead to allow us to
look up the index in `sources_` given the actual string.

Reviewed By: ridiculousfish

Differential Revision: D14530422

fbshipit-source-id: 9441fe8ecccf4e9bf45dfbfc7242d583090c0254