Support multiple segments in source maps.
Summary:
To allow having multiple segment, we use the CJS module offset of a given
BytecodeModule and encode that as the line number in the source map.  The
line at index `0` (the "first" line) will represent the information in
the BytecodeModule representing segment `0`.

We use the `cjsModuleOffset` to avoid modifying the bytecode format,
as the `BytecodeModule` doesn't currently encode the actual segment IDs,
and so the VM has no way of passing that information to `JSError`.

Update the `addMappingsLine` function to accommodate this and allow
skipping to whatever line the user chooses.

Reviewed By: ridiculousfish

Differential Revision: D14530423

fbshipit-source-id: e74d3ec5c7b658fbbd179c59dc2287d558ec2c06