Merge source maps.
Summary:
We have a list of input source maps in `SourceMapGenerator`,
and if they exist, we merge them with the mappings provided within
the `SourceMapGenerator`. This works by creating a new `SourceMapGenerator`
and outputting all the merged mappings into it, then recursing at most
one level.

Note that we account for the lack of an input source map or a relevant
mapping by outputting the original represented line and column,
and simply transferring over the original source file name to the new
`SourceMapGenerator`.

Reviewed By: ridiculousfish

Differential Revision: D14530421

fbshipit-source-id: 878f31c5c42cda960d0af4498a21f837a6f985f7