Add test for UBSAN overflow and blacklist function in dtoa
Summary:
UBSAN found a problem in dtoa's function `g_strtod`, which turns a string into a double.
When the string in question is a very large double (such as `55e5555555555555555`),
it should become the double value `Infinity`. dtoa correctly does this, but assumes 2's complement
signed integer arithmetic in order to do so. This is technically undefined behavior, but
all the platforms we care about have the correct semantics.

Add a blacklist for that function.

Reviewed By: tmikov

Differential Revision: D14603532

fbshipit-source-id: 658058503aa731c2520035eebeaa79a0bcebf015