Make Handle-SAN do full collections on NCGen
Summary:
If Handle-SAN is on, it should cause some sort of heap-moving operation.
In the past, this was implemented as a straight-up heap mover (`swapToFreshHeap`).
This has been un-implemented for a long time, and is not expected to be implemented
anytime soon.

Doing a full collection is pretty much guaranteed to move everything (although not to a fresh heap
space). Not perfect, but better than nothing. And re-allocating a bunch of segments is a lot
slower.

Reviewed By: davedets

Differential Revision: D14656578

fbshipit-source-id: 7c3fdb88ac54cd031a55755a5281a19e6031b08b