Stop object buffer serialization at getter/setter store
Summary:
The bug is when there is a getter or setter in the object literal, the order of the properties
becomes different from the definition. This is because the object buffer lowering pass
assumes every property in the object literal is a `StoreOwnPropertyInst`, so it will ignore
the getter and setters, which is a `StoreGetterSetterInst`, and the getter/setter will be
added after the object is initialized from deserialize from the object buffer.

The fix is to simply stop when we see the first getter/setter store instruction during the
object buffer serialization.

Reviewed By: amnn

Differential Revision: D14674554

fbshipit-source-id: 6366903a3baced6b0098f7c87db8156e8561b2a2