Log fragmentation of bytecode pages in RAM
Summary: Count the number of contiguous runs of bytecode pages in RAM. This metric will be affected by both how well readahead is working, and by how well-ordered our bytecode is w.r.t. our access. In particular, we'd like to know if this number is very different between production and the lab.

Reviewed By: ridiculousfish

Differential Revision: D14686835

fbshipit-source-id: 94d372256b880b9005666629dab9bff5a97dcf6a