Move TracingRuntime function definitions into the cpp file
Summary:
The body of TracingRuntime was almost all inline in the header, but the functions
are called virtually anyway, which prevents any inlining.
API tracing is also not very perf-sensitive as it is a separate build mode right now.
The benefit is that the bodies of these functions can be modified without re-compiling anything
else.

Differential Revision: D14672971

fbshipit-source-id: a12dfd853802e4c3ab24cb0e14b021dfa1977b69