Handle multi config generator properly in CMakeList
Summary:
MSVC (and XCode) are known as multi config generator in CMake terminology. A single invocation of CMake generates all configurations (Debug, MinSizeRel, etc). For such generators, `${CMAKE_BUILD_TYPE}` is meaningless.

This commits changes CMakeList to avoid testing CMAKE_BUILD_TYPE in conditionals. In some cases, it can be replaced with generator-expressions. There are two cases in Hermes configurations where this is not possible.
1. We selectively add files to the project depending on build type. This goes directly against the idea of multi config generator. It is changed so that the files are always included, but may be empty (by adding #ifdef in the C++ files).
2. We change lit test configuration based on cmake build type. These are left as is because no obvious solution is available. The names of the involved CMake varibles are changed to avoid their usage out of this context.

Reviewed By: tmikov

Differential Revision: D14594564

fbshipit-source-id: 0e490adbae8727eb18cf23901d3de73dd50ab2fe